this should do it for pnp closed loop v 0.0.1
edits:
*added gui components
*took out the trigger for recording spikes in NeuroRighter.  This might be added back in to make recording more precise.
*created closedloopexpt class, in two files.  closedloopext is the basic background worker/constructor stuff, closedlooputils contains the stim/recording methods that will probably need to be made more efficient later
*created pnpclosedloop interface for experiments to implement
*created simple example experiment
HELP: the current structure for these experiments is a separate class that gets an instance of it's own wrapper class (closedloopexpt) to run the actual stim/recording methods.  Anyone got an idea on how to do this more efficiently/elegantly/less retarded?

TODO: make interface plug and play so that the user can compile their own experiment and neurorighter separately, and neurorighter can dynamically look up whatever plug ins are available
make recording methods more efficient
possible edit to stimbuffer stuff so that stimulation information can be appended as the stimbuffer runs
make it possible for pnp closed loop module to have a gui in neurorighter (shouldn't be that bad)


git-svn-id: https://neurorighter.googlecode.com/svn/branches/NeuroRighter_ClosedLoop@179 471701b0-e017-11dd-9759-139c9ab14fc6